pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "log-system"
version = "0.1.0"
description = "Log Analysis System with integrated anomaly detection feature"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "torch>=2.7.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]

[tool.hatch.build.targets.wheel]
packages = ["src/log_system"]
